# ParallelEnumerable.GroupJoin/Join output ordering (fixes #1155)