--- a/src/System.Linq.Parallel/System.Linq.Parallel.sln
+++ b/src/System.Linq.Parallel/System.Linq.Parallel.sln
@@ -7,6 +7,11 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "System.Linq.Parallel.Tests"
 		{BE28323E-327A-4E0F-B7F9-16AB7EAB59DD} = {BE28323E-327A-4E0F-B7F9-16AB7EAB59DD}
 	EndProjectSection
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "System.Linq.Parallel.Performance.Tests", "tests\Performance\System.Linq.Parallel.Performance.Tests.csproj", "{11ABE2F8-4FB9-48AC-91AA-D04503059550}"
+	ProjectSection(ProjectDependencies) = postProject
+		{BE28323E-327A-4E0F-B7F9-16AB7EAB59DD} = {BE28323E-327A-4E0F-B7F9-16AB7EAB59DD}
+	EndProjectSection
+EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "System.Linq.Parallel", "src\System.Linq.Parallel.csproj", "{BE28323E-327A-4E0F-B7F9-16AB7EAB59DD}"
 	ProjectSection(ProjectDependencies) = postProject
 		{670405FE-3828-436E-A668-A0F0B88B57A4} = {670405FE-3828-436E-A668-A0F0B88B57A4}
@@ -30,6 +35,10 @@ Global
 		{A7074928-82C3-4739-88FE-9B528977950C}.Debug|Any CPU.Build.0 = netcoreapp-Debug|Any CPU
 		{A7074928-82C3-4739-88FE-9B528977950C}.Release|Any CPU.ActiveCfg = netcoreapp-Release|Any CPU
 		{A7074928-82C3-4739-88FE-9B528977950C}.Release|Any CPU.Build.0 = netcoreapp-Release|Any CPU
+		{11ABE2F8-4FB9-48AC-91AA-D04503059550}.Debug|Any CPU.ActiveCfg = netcoreapp-Debug|Any CPU
+		{11ABE2F8-4FB9-48AC-91AA-D04503059550}.Debug|Any CPU.Build.0 = netcoreapp-Debug|Any CPU
+		{11ABE2F8-4FB9-48AC-91AA-D04503059550}.Release|Any CPU.ActiveCfg = netcoreapp-Release|Any CPU
+		{11ABE2F8-4FB9-48AC-91AA-D04503059550}.Release|Any CPU.Build.0 = netcoreapp-Release|Any CPU
 		{BE28323E-327A-4E0F-B7F9-16AB7EAB59DD}.Debug|Any CPU.ActiveCfg = netcoreapp-Debug|Any CPU
 		{BE28323E-327A-4E0F-B7F9-16AB7EAB59DD}.Debug|Any CPU.Build.0 = netcoreapp-Debug|Any CPU
 		{BE28323E-327A-4E0F-B7F9-16AB7EAB59DD}.Release|Any CPU.ActiveCfg = netcoreapp-Release|Any CPU
@@ -44,6 +53,7 @@ Global
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{A7074928-82C3-4739-88FE-9B528977950C} = {1A2F9F4A-A032-433E-B914-ADD5992BB178}
+		{11ABE2F8-4FB9-48AC-91AA-D04503059550} = {1A2F9F4A-A032-433E-B914-ADD5992BB178}
 		{BE28323E-327A-4E0F-B7F9-16AB7EAB59DD} = {E107E9C1-E893-4E87-987E-04EF0DCEAEFD}
 		{670405FE-3828-436E-A668-A0F0B88B57A4} = {2E666815-2EDB-464B-9DF6-380BF4789AD4}
 	EndGlobalSection