… src/System.Linq.Parallel/src/System/Linq/Parallel/QueryOperators/OrdinalIndexState.cs
… src/System.Linq.Parallel/src/System/Linq/Parallel/QueryOperators/Unary/GroupByQueryOperator.cs
… src/System.Linq.Parallel/tests/Combinatorial/ParallelQueryCombinationTests.cs
… src/System.Linq.Parallel/tests/Performance/Configurations.props
… src/System.Linq.Parallel/tests/Performance/QueryOperators/Perf.GroupBy.cs
… src/System.Linq.Parallel/tests/Performance/QueryOperators/Perf.GroupJoin.cs
… src/System.Linq.Parallel/tests/Performance/QueryOperators/Perf.Join.cs
… src/System.Linq.Parallel/tests/Performance/System.Linq.Parallel.Performance.Tests.csproj
… src/System.Linq.Parallel/tests/QueryOperators/GroupJoinTests.cs
… src/System.Linq.Parallel/tests/QueryOperators/JoinTests.cs
… src/System.Linq.Parallel/tests/System.Linq.Parallel.Tests.csproj