The current implementations of ParallelEnumerable.GroupJoin and Join do not preserve the order of right(inner) elements.

The common implementation, `HashJoinQueryOperatorEnumerator` currently implements a `QueryOperatorEnumerator` with an order key of the same type as the left(outer) query, thus Join operations are unable to take into account the order of the right elements in their output. It also completely ignores the right order key when building a `HashLookup` of the right elements, so order is not preserved for GroupJoins.

The proposed changes attempt to leave the cases where the right query is unordered with a relatively similar performance profile while adding ordering (when appropriate) in the cases where the right query is ordered (fixes #1155).

Join with an unordered left and ordered right results in unordered output and right ordering is completely ignored.
Join with an ordered left and ordered right results in ordered output ordered first by left order key then by right order key. If collisions exist in the ordering of left elements, then the output will likely not be intuitive, as results with the same left element are not guaranteed (and in fact will probably not be) grouped together. (tests **not** added to confirm the ordering of elements with left collisions...)

GroupJoin with an unordered left and ordered right also still produce unordered output. However, the enumerables of right elements are ordered.
GroupJoin with an ordered left and ordered right will produce results ordered first by the left element order key, then by the least right element order key. An empty grouping of right elements is treated as less than any non-empty grouping. (tests added to confirm the ordering of elements with left collisions)