@Clockwork-Muse 2018-03-10
Wow, has it already been 3 years (by about a week) since I posted this?  Time flies!

Good job on solving this in the short time it did take you.

>The proposed changes attempt to leave the cases where the right query is unordered with a relatively similar performance profile while adding ordering (when appropriate) in the cases where the right query is ordered 

Speaking of performance, I don't think we have performance test for PLINQ.  Do you have any comparisons against baseline PLINQ for how this changes things?

>If collisions exist in the ordering of left elements, then the output will likely not be intuitive, as results with the same left element are not guaranteed (and in fact will probably not be) grouped together. (tests not added to confirm the ordering of elements with left collisions...)

Uh, the "intuitive" results I would have expected from a left sequence with duplicate keys (say, `1, 2, 3, 4, 1`) would have been the original input order.  Is that not what's happening/you're guaranteeing?  Or am I misunderstanding something?  As near as I know, the **left** side has always retained input order....