@jswolf19 2018-03-10
> Speaking of performance, I don't think we have performance test for PLINQ. Do you have any comparisons against baseline PLINQ for how this changes things?

I didn't really look into performance testing as I wasn't sure what approach(es) would have meaning (thus the "attempt" ^^;), and am making that claim because the allocations, struct memory copying and such should be similar in the unordered case. I would be interested in adding perf tests, but I would need some guidance on that, as I mentioned in my comments on the issue.

> Uh, the "intuitive" results I would have expected from a left sequence with duplicate keys (say, 1, 2, 3, 4, 1) would have been the original input order. Is that not what's happening/you're guaranteeing? Or am I misunderstanding something? As near as I know, the left side has always retained input order....

I was referring to a `Join` of a left side after an `OrderBy` where multiple elements are equal based on the order key (in your example, the order key is index, so there are collisions for equality checks, but not for ordering). 
For example, 
```
var vals = ParallelEnumerable.Range(0, 4).OrderBy(x => x % 2)
                                             .Join(ParallelEnumerable.Range(0, 16).AsOrdered(),
                                                     x => x, y => y % 4, (x, y) => new KeyValuePair<int, int>(x, y));
Console.WriteLine(string.Join(" ", vals.Select(kv => $"({kv.Key}, {kv.Value})")));
```
outputs (after the fix)
(0, 0) (2, 2) (0, 4) (2, 6) (0, 8) (2, 10) (0, 12) (2, 14) (1, 1) (3, 3) (1, 5) (3, 7) (1, 9) (3, 11) (1, 13) (3, 15)

If my understanding is correct, this should be consistent with the current implementation when only the left input is ordered (the results for 0 and 2 maybe intermingled), but it may not manifest often, if at all, depending on how partition merging is implemented. However, it is exacerbated by the reordering for the right keys.