@Clockwork-Muse 2018-03-10
Ah.  Hrm, I don't know.  On the current version I'm seeing the left items still in initial order, segmented by the ordering criteria (so, 0, 2, 1, 3).  Which is what I would expect; after all, you don't _have_ multiple left elements for each right element, and it's the same left element being repeated (so should stay the same).  Now whether it starts with 0 or 2 is undefined (since equal ordering key), but I would have the same element to be returned, I guess.

@stephentoub - can you give any guidance on expected behavior.