@jswolf19 2018-03-11
@Clockwork-Muse You can see it manifest if you append further operations on.

For example, with the current version, running
```
var vals = ParallelEnumerable.Range(0, 4).OrderBy(x => x % 2)
                             .Join(ParallelEnumerable.Range(0, 16),
                                   x => x, y => y % 4, (x, y) => (x, y))
                             .Join(ParallelEnumerable.Range(0, 16),
                                   t => t.y, z => z, (t, z) => (t.x, t.y, z));
Console.WriteLine(string.Join(" ", vals.Select(t => $"({t.x}, {t.y}, {t.z})")));
```
I got the result
(0, 8, 8) (0, 0, 0) (2, 10, 10) (2, 2, 2) (0, 12, 12) (0, 4, 4) (2, 14, 14) (2, 6, 6) (1, 9, 9) (1, 1, 1) (3, 11, 11) (3, 3, 3) (1, 13, 13) (1, 5, 5) (3, 15, 15) (3, 7, 7)