@Clockwork-Muse 2018-03-11
Behavior confirmed, although mine is giving groups of four, not two (which implies it is partition based).
So we need to figure out if that's the intended behavior, otherwise at least no regression.