@jswolf19 2018-03-11
@Clockwork-Muse yeah, I guess creating examples with the count of elements being a power of 2 could make the undesired behavior harder to observe on some systems... ^^