@jswolf19 2018-03-25
@tarekgh Sorry for not getting back sooner. I was trying to get some time to first element tests, but (I think due to buffering?) it's difficult to get good timings. I am adding a couple of changes to the perf tests. I put my timings up below.

First, I wasn't calculating the right key correctly, so it ended up being 20 lefts with 100 rights and the remaining 80 lefts having no rights. I adjusted the left count to 75, with each left having 20 rights. This results in exaggerating the reduced performance.

Second, I am adding group by tests, as they give a good basis for why group join with an ordered right has such a reduced performance,  since I am using the same logic and `OrderedGroupByGrouping` class for the right enumerable. If I comment out the calls to `DoneAdding` on the `OrderedGroupByGrouping`s, I can roughly half the time of GroupJoin CrossProduct with ordered right. If I also alter the handling of unordered rights to use `GrowingArray`s (which is what `OrderedGroupByGrouping` uses) instead of `ListChunk`s, then I get a similar times and allocations to the ordered rights. This leads me to believe that improving performance of `GroupBy`s `OrderedGroupByGrouping` would be the most productive (and would also improve the performance of `GroupBy`).

## without the changes
```
   System.Linq.Parallel.Performance.Tests.dll                                             | Metric         | Unit  | Iterations |    Average |    STDEV.S |        Min |        Max
  :-------------------------------------------------------------------------------------- |:-------------- |:-----:|:----------:| ----------:| ----------:| ----------:| ----------:
   System.Linq.Parallel.Tests.GroupByPerfTestsOrdered.CrossProduct                        | Duration       | msec  |     14     |    760.695 |     64.646 |    700.242 |    865.756
   System.Linq.Parallel.Tests.GroupByPerfTestsOrdered.CrossProduct                        | GC Allocations | bytes |     14     | 1.252E+010 | 3.520E+008 | 1.174E+010 | 1.284E+010
   System.Linq.Parallel.Tests.GroupByPerfTestsOrdered.QueryCreation                       | Duration       | msec  |     30     |    333.431 |      2.431 |    328.953 |    341.192
   System.Linq.Parallel.Tests.GroupByPerfTestsOrdered.QueryCreation                       | GC Allocations | bytes |     30     | 6.605E+008 |  34286.849 | 6.604E+008 | 6.606E+008
   System.Linq.Parallel.Tests.GroupByPerfTestsUnordered.CrossProduct                      | Duration       | msec  |    100     |     82.922 |      3.138 |     78.875 |    102.261
   System.Linq.Parallel.Tests.GroupByPerfTestsUnordered.CrossProduct                      | GC Allocations | bytes |    100     | 5.481E+008 | 1.067E+007 | 4.952E+008 | 5.699E+008
   System.Linq.Parallel.Tests.GroupByPerfTestsUnordered.QueryCreation                     | Duration       | msec  |     35     |    289.899 |      6.203 |    274.024 |    303.880
   System.Linq.Parallel.Tests.GroupByPerfTestsUnordered.QueryCreation                     | GC Allocations | bytes |     35     | 5.801E+008 |  58620.355 | 5.799E+008 | 5.801E+008
   System.Linq.Parallel.Tests.GroupJoinPerfTestsOrderedLeftOrderedRight.CrossProduct      | Duration       | msec  |     59     |    170.163 |     11.347 |    160.130 |    239.527
   System.Linq.Parallel.Tests.GroupJoinPerfTestsOrderedLeftOrderedRight.CrossProduct      | GC Allocations | bytes |     59     | 1.151E+009 | 3.338E+007 | 1.063E+009 | 1.199E+009
   System.Linq.Parallel.Tests.GroupJoinPerfTestsOrderedLeftOrderedRight.QueryCreation     | Duration       | msec  |     19     |    539.009 |     10.172 |    527.433 |    572.971
   System.Linq.Parallel.Tests.GroupJoinPerfTestsOrderedLeftOrderedRight.QueryCreation     | GC Allocations | bytes |     19     | 9.502E+008 |  45982.312 | 9.502E+008 | 9.503E+008
   System.Linq.Parallel.Tests.GroupJoinPerfTestsOrderedLeftUnorderedRight.CrossProduct    | Duration       | msec  |     76     |    131.558 |      3.569 |    123.823 |    142.796
   System.Linq.Parallel.Tests.GroupJoinPerfTestsOrderedLeftUnorderedRight.CrossProduct    | GC Allocations | bytes |     76     | 1.296E+009 | 1.813E+007 | 1.245E+009 | 1.330E+009
   System.Linq.Parallel.Tests.GroupJoinPerfTestsOrderedLeftUnorderedRight.QueryCreation   | Duration       | msec  |     21     |    483.094 |     10.668 |    467.434 |    516.136
   System.Linq.Parallel.Tests.GroupJoinPerfTestsOrderedLeftUnorderedRight.QueryCreation   | GC Allocations | bytes |     21     | 8.697E+008 |  25308.625 | 8.696E+008 | 8.698E+008
   System.Linq.Parallel.Tests.GroupJoinPerfTestsUnorderedLeftOrderedRight.CrossProduct    | Duration       | msec  |     86     |    117.661 |      6.647 |    108.455 |    143.154
   System.Linq.Parallel.Tests.GroupJoinPerfTestsUnorderedLeftOrderedRight.CrossProduct    | GC Allocations | bytes |     86     | 1.063E+009 | 3.372E+007 | 9.420E+008 | 1.117E+009
   System.Linq.Parallel.Tests.GroupJoinPerfTestsUnorderedLeftOrderedRight.QueryCreation   | Duration       | msec  |     22     |    454.708 |      8.753 |    443.211 |    478.832
   System.Linq.Parallel.Tests.GroupJoinPerfTestsUnorderedLeftOrderedRight.QueryCreation   | GC Allocations | bytes |     22     | 8.697E+008 |  11187.903 | 8.696E+008 | 8.697E+008
   System.Linq.Parallel.Tests.GroupJoinPerfTestsUnorderedLeftUnorderedRight.CrossProduct  | Duration       | msec  |     86     |    116.656 |      3.009 |    110.341 |    128.787
   System.Linq.Parallel.Tests.GroupJoinPerfTestsUnorderedLeftUnorderedRight.CrossProduct  | GC Allocations | bytes |     86     | 1.048E+009 | 1.581E+007 | 9.982E+008 | 1.083E+009
   System.Linq.Parallel.Tests.GroupJoinPerfTestsUnorderedLeftUnorderedRight.QueryCreation | Duration       | msec  |     25     |    409.243 |      6.523 |    395.472 |    428.345
   System.Linq.Parallel.Tests.GroupJoinPerfTestsUnorderedLeftUnorderedRight.QueryCreation | GC Allocations | bytes |     25     | 7.890E+008 |  35733.923 | 7.889E+008 | 7.891E+008
   System.Linq.Parallel.Tests.JoinPerfTestsOrderedLeftOrderedRight.CrossProduct           | Duration       | msec  |     37     |    273.498 |     32.335 |    218.635 |    320.807
   System.Linq.Parallel.Tests.JoinPerfTestsOrderedLeftOrderedRight.CrossProduct           | GC Allocations | bytes |     37     | 1.430E+009 | 1.170E+008 | 1.308E+009 | 1.635E+009
   System.Linq.Parallel.Tests.JoinPerfTestsOrderedLeftOrderedRight.QueryCreation          | Duration       | msec  |     19     |    546.284 |     11.531 |    527.966 |    569.332
   System.Linq.Parallel.Tests.JoinPerfTestsOrderedLeftOrderedRight.QueryCreation          | GC Allocations | bytes |     19     | 9.503E+008 |  51382.809 | 9.502E+008 | 9.504E+008
   System.Linq.Parallel.Tests.JoinPerfTestsOrderedLeftUnorderedRight.CrossProduct         | Duration       | msec  |     43     |    233.106 |     14.772 |    216.160 |    292.105
   System.Linq.Parallel.Tests.JoinPerfTestsOrderedLeftUnorderedRight.CrossProduct         | GC Allocations | bytes |     43     | 1.580E+009 | 5.121E+007 | 1.368E+009 | 1.641E+009
   System.Linq.Parallel.Tests.JoinPerfTestsOrderedLeftUnorderedRight.QueryCreation        | Duration       | msec  |     21     |    488.808 |     16.483 |    465.448 |    521.747
   System.Linq.Parallel.Tests.JoinPerfTestsOrderedLeftUnorderedRight.QueryCreation        | GC Allocations | bytes |     21     | 8.698E+008 |  65300.305 | 8.697E+008 | 8.700E+008
   System.Linq.Parallel.Tests.JoinPerfTestsUnorderedLeftOrderedRight.CrossProduct         | Duration       | msec  |     47     |    212.896 |      9.041 |    201.699 |    257.857
   System.Linq.Parallel.Tests.JoinPerfTestsUnorderedLeftOrderedRight.CrossProduct         | GC Allocations | bytes |     47     | 8.604E+008 | 2.516E+007 | 8.123E+008 | 9.422E+008
   System.Linq.Parallel.Tests.JoinPerfTestsUnorderedLeftOrderedRight.QueryCreation        | Duration       | msec  |     21     |    491.339 |     15.702 |    468.901 |    523.348
   System.Linq.Parallel.Tests.JoinPerfTestsUnorderedLeftOrderedRight.QueryCreation        | GC Allocations | bytes |     21     | 8.699E+008 |  26520.601 | 8.698E+008 | 8.700E+008
   System.Linq.Parallel.Tests.JoinPerfTestsUnorderedLeftUnorderedRight.CrossProduct       | Duration       | msec  |     53     |    190.740 |     17.166 |    176.492 |    248.683
   System.Linq.Parallel.Tests.JoinPerfTestsUnorderedLeftUnorderedRight.CrossProduct       | GC Allocations | bytes |     53     | 1.035E+009 | 8.723E+007 | 7.287E+008 | 1.126E+009
   System.Linq.Parallel.Tests.JoinPerfTestsUnorderedLeftUnorderedRight.QueryCreation      | Duration       | msec  |     24     |    433.543 |     16.968 |    408.515 |    460.216
   System.Linq.Parallel.Tests.JoinPerfTestsUnorderedLeftUnorderedRight.QueryCreation      | GC Allocations | bytes |     24     | 7.892E+008 |  46283.311 | 7.891E+008 | 7.893E+008
```

## with the changes
```
   System.Linq.Parallel.Performance.Tests.dll                                             | Metric         | Unit  | Iterations |    Average |    STDEV.S |        Min |        Max
  :-------------------------------------------------------------------------------------- |:-------------- |:-----:|:----------:| ----------:| ----------:| ----------:| ----------:
   System.Linq.Parallel.Tests.GroupByPerfTestsOrdered.CrossProduct                        | Duration       | msec  |     13     |    769.787 |     73.219 |    702.363 |    926.106
   System.Linq.Parallel.Tests.GroupByPerfTestsOrdered.CrossProduct                        | GC Allocations | bytes |     13     | 1.257E+010 | 2.145E+008 | 1.213E+010 | 1.284E+010
   System.Linq.Parallel.Tests.GroupByPerfTestsOrdered.QueryCreation                       | Duration       | msec  |     30     |    333.448 |      7.812 |    324.182 |    366.332
   System.Linq.Parallel.Tests.GroupByPerfTestsOrdered.QueryCreation                       | GC Allocations | bytes |     30     | 6.606E+008 |  40729.674 | 6.605E+008 | 6.606E+008
   System.Linq.Parallel.Tests.GroupByPerfTestsUnordered.CrossProduct                      | Duration       | msec  |    100     |     83.377 |      5.956 |     77.053 |    110.172
   System.Linq.Parallel.Tests.GroupByPerfTestsUnordered.CrossProduct                      | GC Allocations | bytes |    100     | 5.401E+008 | 1.638E+007 | 4.725E+008 | 5.650E+008
   System.Linq.Parallel.Tests.GroupByPerfTestsUnordered.QueryCreation                     | Duration       | msec  |     36     |    280.636 |      6.901 |    270.010 |    304.387
   System.Linq.Parallel.Tests.GroupByPerfTestsUnordered.QueryCreation                     | GC Allocations | bytes |     36     | 5.801E+008 |  85908.770 | 5.797E+008 | 5.802E+008
   System.Linq.Parallel.Tests.GroupJoinPerfTestsOrderedLeftOrderedRight.CrossProduct      | Duration       | msec  |     13     |    806.490 |     27.541 |    775.169 |    864.054
   System.Linq.Parallel.Tests.GroupJoinPerfTestsOrderedLeftOrderedRight.CrossProduct      | GC Allocations | bytes |     13     | 1.353E+010 | 2.108E+008 | 1.314E+010 | 1.378E+010
   System.Linq.Parallel.Tests.GroupJoinPerfTestsOrderedLeftOrderedRight.QueryCreation     | Duration       | msec  |     20     |    518.815 |     17.918 |    492.658 |    562.566
   System.Linq.Parallel.Tests.GroupJoinPerfTestsOrderedLeftOrderedRight.QueryCreation     | GC Allocations | bytes |     20     | 9.503E+008 |  34741.025 | 9.502E+008 | 9.503E+008
   System.Linq.Parallel.Tests.GroupJoinPerfTestsOrderedLeftUnorderedRight.CrossProduct    | Duration       | msec  |     69     |    146.242 |     22.461 |    120.940 |    197.178
   System.Linq.Parallel.Tests.GroupJoinPerfTestsOrderedLeftUnorderedRight.CrossProduct    | GC Allocations | bytes |     69     | 1.273E+009 | 7.667E+007 | 9.369E+008 | 1.374E+009
   System.Linq.Parallel.Tests.GroupJoinPerfTestsOrderedLeftUnorderedRight.QueryCreation   | Duration       | msec  |     22     |    471.280 |     20.785 |    452.752 |    520.906
   System.Linq.Parallel.Tests.GroupJoinPerfTestsOrderedLeftUnorderedRight.QueryCreation   | GC Allocations | bytes |     22     | 8.697E+008 |  63540.887 | 8.697E+008 | 8.700E+008
   System.Linq.Parallel.Tests.GroupJoinPerfTestsUnorderedLeftOrderedRight.CrossProduct    | Duration       | msec  |     14     |    741.013 |     12.925 |    725.544 |    774.798
   System.Linq.Parallel.Tests.GroupJoinPerfTestsUnorderedLeftOrderedRight.CrossProduct    | GC Allocations | bytes |     14     | 1.349E+010 | 7.443E+007 | 1.331E+010 | 1.358E+010
   System.Linq.Parallel.Tests.GroupJoinPerfTestsUnorderedLeftOrderedRight.QueryCreation   | Duration       | msec  |     22     |    460.891 |      8.778 |    444.077 |    481.303
   System.Linq.Parallel.Tests.GroupJoinPerfTestsUnorderedLeftOrderedRight.QueryCreation   | GC Allocations | bytes |     22     | 8.697E+008 |  49997.271 | 8.696E+008 | 8.698E+008
   System.Linq.Parallel.Tests.GroupJoinPerfTestsUnorderedLeftUnorderedRight.CrossProduct  | Duration       | msec  |     88     |    114.575 |      3.374 |    107.939 |    126.663
   System.Linq.Parallel.Tests.GroupJoinPerfTestsUnorderedLeftUnorderedRight.CrossProduct  | GC Allocations | bytes |     88     | 1.084E+009 | 2.456E+007 | 9.940E+008 | 1.135E+009
   System.Linq.Parallel.Tests.GroupJoinPerfTestsUnorderedLeftUnorderedRight.QueryCreation | Duration       | msec  |     25     |    412.047 |      7.850 |    392.589 |    429.919
   System.Linq.Parallel.Tests.GroupJoinPerfTestsUnorderedLeftUnorderedRight.QueryCreation | GC Allocations | bytes |     25     | 7.892E+008 |  58794.915 | 7.891E+008 | 7.893E+008
   System.Linq.Parallel.Tests.JoinPerfTestsOrderedLeftOrderedRight.CrossProduct           | Duration       | msec  |     37     |    275.109 |      3.405 |    269.940 |    283.171
   System.Linq.Parallel.Tests.JoinPerfTestsOrderedLeftOrderedRight.CrossProduct           | GC Allocations | bytes |     37     | 1.781E+009 | 1.287E+007 | 1.751E+009 | 1.807E+009
   System.Linq.Parallel.Tests.JoinPerfTestsOrderedLeftOrderedRight.QueryCreation          | Duration       | msec  |     20     |    520.143 |     11.576 |    510.094 |    563.131
   System.Linq.Parallel.Tests.JoinPerfTestsOrderedLeftOrderedRight.QueryCreation          | GC Allocations | bytes |     20     | 9.502E+008 |  43197.139 | 9.502E+008 | 9.504E+008
   System.Linq.Parallel.Tests.JoinPerfTestsOrderedLeftUnorderedRight.CrossProduct         | Duration       | msec  |     45     |    223.301 |      4.736 |    212.746 |    234.239
   System.Linq.Parallel.Tests.JoinPerfTestsOrderedLeftUnorderedRight.CrossProduct         | GC Allocations | bytes |     45     | 1.619E+009 | 2.292E+007 | 1.578E+009 | 1.660E+009
   System.Linq.Parallel.Tests.JoinPerfTestsOrderedLeftUnorderedRight.QueryCreation        | Duration       | msec  |     22     |    464.217 |      4.103 |    457.233 |    474.279
   System.Linq.Parallel.Tests.JoinPerfTestsOrderedLeftUnorderedRight.QueryCreation        | GC Allocations | bytes |     22     | 8.697E+008 |  31624.775 | 8.696E+008 | 8.698E+008
   System.Linq.Parallel.Tests.JoinPerfTestsUnorderedLeftOrderedRight.CrossProduct         | Duration       | msec  |     58     |    174.503 |      3.564 |    167.686 |    186.790
   System.Linq.Parallel.Tests.JoinPerfTestsUnorderedLeftOrderedRight.CrossProduct         | GC Allocations | bytes |     58     | 1.187E+009 | 6.448E+007 | 1.038E+009 | 1.275E+009
   System.Linq.Parallel.Tests.JoinPerfTestsUnorderedLeftOrderedRight.QueryCreation        | Duration       | msec  |     22     |    463.578 |      1.843 |    460.301 |    466.649
   System.Linq.Parallel.Tests.JoinPerfTestsUnorderedLeftOrderedRight.QueryCreation        | GC Allocations | bytes |     22     | 8.697E+008 |  44186.754 | 8.696E+008 | 8.698E+008
   System.Linq.Parallel.Tests.JoinPerfTestsUnorderedLeftUnorderedRight.CrossProduct       | Duration       | msec  |     57     |    177.353 |      5.923 |    171.343 |    216.319
   System.Linq.Parallel.Tests.JoinPerfTestsUnorderedLeftUnorderedRight.CrossProduct       | GC Allocations | bytes |     57     | 1.232E+009 | 3.381E+007 | 1.146E+009 | 1.296E+009
   System.Linq.Parallel.Tests.JoinPerfTestsUnorderedLeftUnorderedRight.QueryCreation      | Duration       | msec  |     25     |    404.643 |      4.475 |    396.462 |    418.647
   System.Linq.Parallel.Tests.JoinPerfTestsUnorderedLeftUnorderedRight.QueryCreation      | GC Allocations | bytes |     25     | 7.892E+008 |  64316.955 | 7.889E+008 | 7.893E+008
```